Changed datatype of X to float32

+++ environment.py
@@ -13,7 +13,7 @@ tf.compat.v1.enable_v2_behavior()
 class ClassifyEnv(py_environment.PyEnvironment):
     def __init__(self, X_train, y_train, imb_rate):
         self._action_spec = array_spec.BoundedArraySpec(shape=(), dtype=np.int32, minimum=0, maximum=1, name='action')
-        self._observation_spec = array_spec.ArraySpec(shape=(29,), dtype=np.float64, name='observation')
+        self._observation_spec = array_spec.ArraySpec(shape=(29,), dtype=np.float32, name='observation')
         self._episode_ended = False
 
         self.X_train = X_train
